+++ app/renderer/stores/task_store.js
@@ -1,6 +1,7 @@
 "use strict";
 
 const Suite = require('../../common/suite');
+const Task = require('../../common/task');
 const TasksHandler = require('../tasks_handler');
 const TaskImporter = require('../../common/task_importer');
 
@@ -69,9 +70,12 @@ module.exports = {
             const suite = state.suites[data.suite];
             suite.title = data.title;
         },
-        updateTasks(state, data) {
+        updateSuiteTasks(state, data) {
             const suite = state.suites[data.suite];
             suite.tasks = data.tasks;
+        },
+        updateSuites(state, data) {
+            state.suites = data;
         },
         validateTaskName(state, data) {
             const suite = state.suites[data.suite];
@@ -159,6 +163,14 @@ module.exports = {
         deleteSuite(context, i) {
             context.dispatch("stopSuite", i);
             context.commit("_deleteSuite", i);
+        },
+        duplicateTask(context, data) {
+            const oldTask = context.getters.suites[data.suite].getTask(data.task);
+            const newTask = new Task(oldTask.title, oldTask.path, oldTask.command);
+            context.commit("addTask", {
+                index: data.suite,
+                task: newTask
+            });
         }
     }
 };