+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.5.2 / ####-##-##
+==================
+
+  * Suites can now be reordered
+  * Run/Stop option in context menu
+  * Duplicate task option in context menu
+  * Fixed delete suite bug
+
 0.5.1 / 2018-06-14
 ==================